chore: move routes
